修复工具整数参数系统性失效 + 图片编解码边界 + 渲染乱码 8 项
- ToolArgs.GetInt 统一兼容 int/long/double/string，替换 19 处 is int 判断（JSON 整数解析为 long 导致参数静默丢参回退默认）
- MultiEditTool.ParseEdits 补 IEnumerable 分支（List<object?> 形态，工具此前完全失效）
- DrawCommands path 首点坐标丢失（cx 初始应为 NaN）
- PngDecoder chunk 长度整数溢出绕过越界检查（改 long 比较）
- BmpCodec 32 位 BI_RGB 保留字节误当 alpha（图像全透明）
- /history 预览 Substring 代理对切半（提取 BuildHistoryPreview 对齐）
- /jobs 命令列表 [..57] 代理对切半（改 TruncateByRunes）

+++ WayCoder/Infra/PngDecoder.cs
@@ -33,7 +33,7 @@ public static class PngDecoder
             if (len < 0) throw new FormatException("PNG chunk 长度非法"); // 负数长度会让 off 回退，造成死循环
             string type = Encoding.ASCII.GetString(data, off + 4, 4);
             off += 8;
-            if (off + len > data.Length) throw new FormatException("PNG chunk 越界");
+            if ((long)off + len > data.Length) throw new FormatException("PNG chunk 越界"); // 用 long 防 len=int.MaxValue 时 off+len 溢出为负、绕过检查
             switch (type)
             {
                 case "IHDR":
